#12816 - Cleanup TreeNodes
