+++ primefaces/src/main/java/org/primefaces/model/DefaultTreeNode.java
@@ -33,21 +33,21 @@ public class DefaultTreeNode<T> implements TreeNode<T>, Serializable {
     private static final long serialVersionUID = 1L;
 
 
-    private String type;
+    protected String type;
 
-    private T data;
+    protected T data;
 
-    private List<TreeNode<T>> children;
+    protected TreeNodeChildren<T> children;
 
-    private TreeNode parent;
+    protected TreeNode parent;
 
-    private boolean expanded;
+    protected boolean expanded;
 
-    private boolean selected;
+    protected boolean selected;
 
-    private boolean selectable = true;
+    protected boolean selectable = true;
 
-    private String rowKey;
+    protected String rowKey;
 
     public DefaultTreeNode() {
         this(null);
@@ -70,8 +70,8 @@ public class DefaultTreeNode<T> implements TreeNode<T>, Serializable {
         }
     }
 
-    protected List<TreeNode<T>> initChildren() {
-        return new TreeNodeChildren(this);
+    protected TreeNodeChildren<T> initChildren() {
+        return new DefaultTreeNodeChildren<>(this);
     }
 
     @Override
@@ -94,18 +94,13 @@ public class DefaultTreeNode<T> implements TreeNode<T>, Serializable {
     }
 
     @Override
-    public List<TreeNode<T>> getChildren() {
+    public TreeNodeChildren<T> getChildren() {
         return children;
     }
 
     public void setChildren(List<TreeNode<T>> children) {
-        if (children instanceof TreeNodeChildren) {
-            this.children = children;
-        }
-        else {
-            this.children = new TreeNodeChildren(this);
-            this.children.addAll(children);
-        }
+        this.children = initChildren();
+        this.children.addAll(children);
     }
 
     @Override
